Refactor: Select return type
The ret type in the select method of the DBInteractor class was a string vec. This just returned the whole table without rows seperated. Included a temp vec that is pushed to payload so now each row is its own vec within the vec "payload". [Vec<String> --> Vec<Vec<String> :)]